major_webapp/requirement.txt
asgiref==3.2.10
blis==0.4.1
certifi==2020.11.8
chardet==3.0.4
cymem==2.0.4
decorator==4.4.2
Django==3.1.1
django-cors-headers==3.5.0
django-crontab==0.7.1
django-extensions==3.0.9
djangorestframework==3.11.1
feedparser==6.0.2
idna==2.10
importlib-metadata==3.1.0
joblib==0.17.0
murmurhash==1.0.4
networkx==2.5
numpy==1.19.4
pandas==1.1.4
plac==1.1.3
preshed==3.0.4
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
sklearn-hierarchical-classification==1.3.2
spacy==3.0.3
sqlparse==0.3.1
threadpoolctl==2.1.0
tqdm==4.53.0
urllib3==1.26.2
zipp==3.4.0
